refactor(scenario): split trim-past-UT cluster into ParsekScenario.Trim.cs (flip partial)
