+++ Source/Parsek/ParsekScenario.cs
@@ -29,7 +29,7 @@ namespace Parsek
     /// </summary>
     [KSPScenario(ScenarioCreationOptions.AddToAllGames,
         GameScenes.FLIGHT, GameScenes.SPACECENTER, GameScenes.TRACKSTATION, GameScenes.EDITOR)]
-    public class ParsekScenario : ScenarioModule
+    public partial class ParsekScenario : ScenarioModule
     {
         internal const string RewindSpawnSuppressionReasonSameRecording = "same-recording";
         internal const string RewindSpawnSuppressionReasonSameTreeFutureRecording = "same-tree-future-recording";
@@ -4868,401 +4868,6 @@ namespace Parsek
                     $"suppression but no active tree node was present reason='{noActiveSuppressReason}'");
             }
             return false;
-        }
-
-        /// <summary>
-        /// Truncates a live active-tree recording to the current quickload-resume UT so
-        /// post-load recording can continue from the restored timeline without appending
-        /// stale samples/events from the pre-load future. Returns true when any payload was
-        /// removed or clipped.
-        /// </summary>
-        internal static bool TrimRecordingPastUT(Recording rec, double cutoffUT)
-        {
-            if (rec == null || double.IsNaN(cutoffUT) || double.IsInfinity(cutoffUT))
-                return false;
-
-            bool mutated = false;
-
-            mutated |= RemoveItemsPastUT(rec.Points, cutoffUT, p => p.ut);
-            mutated |= TrimOrbitSegmentsPastUT(rec.OrbitSegments, cutoffUT);
-            mutated |= RemoveItemsPastUT(rec.PartEvents, cutoffUT, e => e.ut);
-            mutated |= RemoveItemsPastUT(rec.FlagEvents, cutoffUT, e => e.ut);
-            mutated |= RemoveItemsPastUT(rec.SegmentEvents, cutoffUT, e => e.ut);
-            mutated |= TrimTrackSectionsPastUT(rec.TrackSections, cutoffUT);
-
-            if (!double.IsNaN(rec.ExplicitStartUT) && rec.ExplicitStartUT > cutoffUT)
-            {
-                rec.ExplicitStartUT = cutoffUT;
-                mutated = true;
-            }
-
-            if (double.IsNaN(rec.ExplicitEndUT) || rec.ExplicitEndUT > cutoffUT)
-            {
-                rec.ExplicitEndUT = cutoffUT;
-                mutated = true;
-            }
-
-            if (mutated)
-                rec.MarkFilesDirty();
-
-            return mutated;
-        }
-
-        internal static bool TrimRecordingTreePastUT(RecordingTree tree, double cutoffUT)
-        {
-            if (tree == null || tree.Recordings == null || tree.Recordings.Count == 0
-                || double.IsNaN(cutoffUT) || double.IsInfinity(cutoffUT))
-            {
-                return false;
-            }
-
-            bool mutated = false;
-            int trimmedCount = 0;
-            HashSet<string> futureOnlyIds = CollectFutureOnlyRecordingIds(tree, cutoffUT);
-            int recordingCountBeforeTrim = tree.Recordings.Count;
-            foreach (Recording rec in tree.Recordings.Values)
-            {
-                if (TrimRecordingPastUT(rec, cutoffUT))
-                {
-                    mutated = true;
-                    trimmedCount++;
-                }
-            }
-
-            if (mutated || (futureOnlyIds != null && futureOnlyIds.Count > 0))
-            {
-                int prunedRecordings = PruneFutureOnlyRecordings(tree, futureOnlyIds);
-                int prunedBranchPoints = RemoveEmptyBranchPoints(tree);
-                if (prunedRecordings > 0 || prunedBranchPoints > 0)
-                    mutated = true;
-
-                tree.RebuildBackgroundMap();
-                ParsekLog.Info("Scenario",
-                    $"Quickload tree trim: tree='{tree.TreeName}' cutoffUT={cutoffUT.ToString("F2", CultureInfo.InvariantCulture)} " +
-                    $"trimmedRecordings={trimmedCount}/{recordingCountBeforeTrim} " +
-                    $"prunedFutureRecordings={prunedRecordings} prunedBranchPoints={prunedBranchPoints} " +
-                    $"backgroundEntries={tree.BackgroundMap.Count}");
-            }
-
-            return mutated;
-        }
-
-        /// <summary>
-        /// Bug #610: scope of the quickload-resume tail trim. Tree-wide is correct
-        /// for F9 quickload — the world rewound, every recording's post-cutoff data
-        /// is stale and future-only recordings never existed at the resume UT.
-        /// Re-Fly is different: the splice has already restored post-RP recordings
-        /// that represent OTHER vessels' continued timelines and the re-flown
-        /// vessel's destroyed-fork; tree-wide trimming would clip and prune them.
-        /// Only the in-place continuation target (the active rec) needs its tail
-        /// trimmed so the recorder can append fresh post-cutoff data without
-        /// colliding with the pre-cutoff timeline.
-        /// </summary>
-        internal enum QuickloadTrimScope
-        {
-            TreeWide = 0,
-            ActiveRecOnly = 1,
-        }
-
-        /// <summary>
-        /// Picks the trim scope based on whether an active Re-Fly session pins
-        /// this tree. Pure function so the decision is unit-testable. The
-        /// <paramref name="reason"/> string is appended to the resume-prep log
-        /// line so the chosen branch is auditable from KSP.log alone (#610).
-        /// </summary>
-        internal static QuickloadTrimScope ChooseQuickloadTrimScope(
-            string treeId,
-            ReFlySessionMarker marker,
-            out string reason)
-        {
-            if (marker == null)
-            {
-                reason = "no-active-refly-marker";
-                return QuickloadTrimScope.TreeWide;
-            }
-            if (string.IsNullOrEmpty(marker.TreeId))
-            {
-                reason = $"refly-marker-has-no-treeid sess={marker.SessionId ?? "<no-id>"}";
-                return QuickloadTrimScope.TreeWide;
-            }
-            if (string.IsNullOrEmpty(treeId))
-            {
-                reason = $"resume-tree-has-no-id markerTree={marker.TreeId}";
-                return QuickloadTrimScope.TreeWide;
-            }
-            if (!string.Equals(marker.TreeId, treeId, StringComparison.Ordinal))
-            {
-                reason = $"refly-marker-tree-mismatch markerTree={marker.TreeId} resumeTree={treeId} sess={marker.SessionId ?? "<no-id>"}";
-                return QuickloadTrimScope.TreeWide;
-            }
-            reason = $"refly-active sess={marker.SessionId ?? "<no-id>"} markerTree={marker.TreeId} originRec={marker.OriginChildRecordingId ?? "<null>"}";
-            return QuickloadTrimScope.ActiveRecOnly;
-        }
-
-        private static HashSet<string> CollectFutureOnlyRecordingIds(RecordingTree tree, double cutoffUT)
-        {
-            HashSet<string> futureOnlyIds = null;
-            foreach (KeyValuePair<string, Recording> kvp in tree.Recordings)
-            {
-                string recordingId = kvp.Key;
-                Recording rec = kvp.Value;
-                if (rec == null
-                    || rec.SidecarLoadFailed
-                    || string.Equals(recordingId, tree.ActiveRecordingId, StringComparison.Ordinal)
-                    || string.Equals(recordingId, tree.RootRecordingId, StringComparison.Ordinal))
-                {
-                    continue;
-                }
-
-                if (rec.StartUT >= cutoffUT)
-                {
-                    if (futureOnlyIds == null)
-                        futureOnlyIds = new HashSet<string>(StringComparer.Ordinal);
-                    futureOnlyIds.Add(recordingId);
-                }
-            }
-
-            return futureOnlyIds;
-        }
-
-        private static int PruneFutureOnlyRecordings(RecordingTree tree, HashSet<string> futureOnlyIds)
-        {
-            if (tree == null || futureOnlyIds == null || futureOnlyIds.Count == 0)
-                return 0;
-
-            int removed = 0;
-            foreach (string recordingId in futureOnlyIds)
-            {
-                if (tree.Recordings.Remove(recordingId))
-                    removed++;
-            }
-
-            if (removed == 0)
-                return 0;
-
-            if (tree.BranchPoints != null)
-            {
-                for (int i = 0; i < tree.BranchPoints.Count; i++)
-                {
-                    BranchPoint bp = tree.BranchPoints[i];
-                    bp.ParentRecordingIds?.RemoveAll(id => futureOnlyIds.Contains(id));
-                    bp.ChildRecordingIds?.RemoveAll(id => futureOnlyIds.Contains(id));
-                }
-            }
-
-            foreach (Recording rec in tree.Recordings.Values)
-            {
-                if (rec != null && futureOnlyIds.Contains(rec.ParentRecordingId))
-                    rec.ParentRecordingId = null;
-            }
-
-            return removed;
-        }
-
-        private static int RemoveEmptyBranchPoints(RecordingTree tree)
-        {
-            if (tree == null || tree.BranchPoints == null || tree.BranchPoints.Count == 0)
-                return 0;
-
-            HashSet<string> removedBranchPointIds = null;
-            int removed = 0;
-            for (int i = tree.BranchPoints.Count - 1; i >= 0; i--)
-            {
-                BranchPoint bp = tree.BranchPoints[i];
-                int parentCount = bp.ParentRecordingIds != null ? bp.ParentRecordingIds.Count : 0;
-                int childCount = bp.ChildRecordingIds != null ? bp.ChildRecordingIds.Count : 0;
-                if (parentCount > 0 && childCount > 0)
-                    continue;
-
-                if (removedBranchPointIds == null)
-                    removedBranchPointIds = new HashSet<string>(StringComparer.Ordinal);
-                removedBranchPointIds.Add(bp.Id);
-                tree.BranchPoints.RemoveAt(i);
-                removed++;
-            }
-
-            if (removedBranchPointIds == null || removedBranchPointIds.Count == 0)
-                return 0;
-
-            foreach (Recording rec in tree.Recordings.Values)
-            {
-                if (rec == null)
-                    continue;
-
-                if (!string.IsNullOrEmpty(rec.ParentBranchPointId)
-                    && removedBranchPointIds.Contains(rec.ParentBranchPointId))
-                {
-                    rec.ParentBranchPointId = null;
-                }
-
-                if (!string.IsNullOrEmpty(rec.ChildBranchPointId)
-                    && removedBranchPointIds.Contains(rec.ChildBranchPointId))
-                {
-                    rec.ChildBranchPointId = null;
-                }
-            }
-
-            return removed;
-        }
-
-        private static bool TrimOrbitSegmentsPastUT(List<OrbitSegment> segments, double cutoffUT)
-        {
-            if (segments == null || segments.Count == 0)
-                return false;
-
-            bool mutated = false;
-            for (int i = segments.Count - 1; i >= 0; i--)
-            {
-                OrbitSegment seg = segments[i];
-                if (seg.startUT >= cutoffUT)
-                {
-                    segments.RemoveAt(i);
-                    mutated = true;
-                    continue;
-                }
-
-                if (seg.endUT > cutoffUT)
-                {
-                    seg.endUT = cutoffUT;
-                    segments[i] = seg;
-                    mutated = true;
-                }
-            }
-
-            return mutated;
-        }
-
-        private static bool TrimTrackSectionsPastUT(List<TrackSection> sections, double cutoffUT)
-        {
-            if (sections == null || sections.Count == 0)
-                return false;
-
-            bool mutated = false;
-            for (int i = sections.Count - 1; i >= 0; i--)
-            {
-                TrackSection section = sections[i];
-                if (section.startUT >= cutoffUT)
-                {
-                    sections.RemoveAt(i);
-                    mutated = true;
-                    continue;
-                }
-
-                bool sectionMutated = false;
-                if (section.endUT > cutoffUT)
-                {
-                    section.endUT = cutoffUT;
-                    sectionMutated = true;
-                }
-
-                sectionMutated |= TrimTrackSectionFramesPastUT(ref section, cutoffUT);
-                sectionMutated |= TrimTrackSectionCheckpointsPastUT(ref section, cutoffUT);
-
-                bool hasFrames = section.frames != null && section.frames.Count > 0;
-                bool hasCheckpoints = section.checkpoints != null && section.checkpoints.Count > 0;
-                if (!hasFrames && !hasCheckpoints)
-                {
-                    sections.RemoveAt(i);
-                    mutated = true;
-                    continue;
-                }
-
-                if (sectionMutated)
-                {
-                    RecomputeTrimmedTrackSectionMetadata(ref section);
-                    sections[i] = section;
-                    mutated = true;
-                }
-            }
-
-            return mutated;
-        }
-
-        private static bool TrimTrackSectionFramesPastUT(ref TrackSection section, double cutoffUT)
-        {
-            if (section.frames == null || section.frames.Count == 0)
-                return false;
-
-            int originalCount = section.frames.Count;
-            for (int i = section.frames.Count - 1; i >= 0; i--)
-            {
-                if (section.frames[i].ut > cutoffUT)
-                    section.frames.RemoveAt(i);
-            }
-
-            if (section.frames.Count == 0)
-                section.frames = null;
-
-            int remainingCount = section.frames != null ? section.frames.Count : 0;
-            return remainingCount != originalCount;
-        }
-
-        private static bool TrimTrackSectionCheckpointsPastUT(ref TrackSection section, double cutoffUT)
-        {
-            if (section.checkpoints == null || section.checkpoints.Count == 0)
-                return false;
-
-            bool mutated = false;
-            for (int i = section.checkpoints.Count - 1; i >= 0; i--)
-            {
-                OrbitSegment checkpoint = section.checkpoints[i];
-                if (checkpoint.startUT >= cutoffUT)
-                {
-                    section.checkpoints.RemoveAt(i);
-                    mutated = true;
-                    continue;
-                }
-
-                if (checkpoint.endUT > cutoffUT)
-                {
-                    checkpoint.endUT = cutoffUT;
-                    section.checkpoints[i] = checkpoint;
-                    mutated = true;
-                }
-            }
-
-            if (section.checkpoints.Count == 0)
-                section.checkpoints = null;
-
-            return mutated;
-        }
-
-        private static void RecomputeTrimmedTrackSectionMetadata(ref TrackSection section)
-        {
-            section.sampleRateHz = 0f;
-            section.minAltitude = float.NaN;
-            section.maxAltitude = float.NaN;
-
-            if (section.frames == null || section.frames.Count == 0)
-                return;
-
-            for (int i = 0; i < section.frames.Count; i++)
-            {
-                float alt = (float)section.frames[i].altitude;
-                if (float.IsNaN(section.minAltitude) || alt < section.minAltitude)
-                    section.minAltitude = alt;
-                if (float.IsNaN(section.maxAltitude) || alt > section.maxAltitude)
-                    section.maxAltitude = alt;
-            }
-
-            double duration = section.endUT - section.startUT;
-            if (duration > 0.0 && section.frames.Count > 1)
-                section.sampleRateHz = (float)(section.frames.Count / duration);
-        }
-
-        private static bool RemoveItemsPastUT<T>(List<T> items, double cutoffUT, Func<T, double> getUT)
-        {
-            if (items == null || items.Count == 0)
-                return false;
-
-            int originalCount = items.Count;
-            for (int i = items.Count - 1; i >= 0; i--)
-            {
-                if (getUT(items[i]) > cutoffUT)
-                    items.RemoveAt(i);
-            }
-
-            return items.Count != originalCount;
         }
 
         internal static bool ShouldKeepPendingTreeAfterHydrationFailure(
